Azure_RAG/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RAG Application Backend with Entity Extraction and Multi-Strategy Search"
readme = "README.md"
authors = [
    { name = "MinhQuanBuiSco", email = "quanbm2710@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "azure-storage-blob>=12.24.0",
    "azure-search-documents>=11.6.0",
    "azure-identity>=1.19.0",
    "openai>=1.58.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf2>=3.0.1",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]
